errand-sequencer/requirements.txt
# Core app
streamlit>=1.28.0
ollama>=0.3.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Agent + tool calling (Ollama)
langchain-core>=0.3.0
langchain-ollama>=0.2.0

# RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# APIs (optional client libs — we use httpx for Google REST)
# googlemaps

# Dev
# pytest>=7.0.0
# ruff
